remote/requirements.txt
gpiozero
requests
python-dotenv